requirements.txt
pytorch
numpy
pandas
scipy
resampy
six
pickle
h5py
librosa

